renamed, fixes #8
