+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v 0.0.4 (2017/10/03)
+* RENAMED npm package to "ngx-youtube-player"
+
+## v 0.0.3 (2017/05/15)
+* updated to work with ng4
+* auto setting for youtube player api protocol (http or https)
+* using angular-library-starter
+* added support for "http" method
+
+## v 0.0.2 (2017/03/29)
+* added examples
+
 ## v 0.0.1 (2017/05/10)
 
 ### Updates
