+++ index.ts
@@ -1,1 +1,1 @@
-import { YoutubePlayerModule } from './src/ng2-youtube-player';
+import { YoutubePlayerModule } from './src/ngx-youtube-player';
