--- src/ng2-youtube-player.ts
@@ -1,7 +1,0 @@
-// Public classes.
-export { YoutubePlayerModule } from './modules/ng2-youtube-player.module';
-export { YoutubePlayerComponent } from './modules/youtube-player.component';
-
-export {
-  IPlayerApiScriptOptions, IPlayerOutputs, IPlayerSize
-} from './models';
