+++ src/ngx-youtube-player.ts
@@ -1,0 +1,9 @@
+// Public classes.
+export {
+  IPlayerApiScriptOptions, IPlayerOutputs, IPlayerSize
+} from './models';
+
+export { YoutubePlayerModule } from './modules/ngx-youtube-player.module';
+export { YoutubePlayerComponent } from './modules/youtube-player.component';
+
+export { YoutubePlayerService } from './services/youtube-player.service';
